Allow ray to auto-detect object store size.
It seems to do a good enough job at this.

Though if this code is using the object store size as a way to
limit the replay buffer size, that's going to become an issue.
When running locally on a small machine, we need to rely on
the object store's spill (page-to-disk) functionality, so we can't
rely on the in-memory size of the object store to control the
replay buffer size.