Add saving and loading of ReplayBuffer as well.
The ReplayBuffer is vital when learning. Losing the replay buffer
causes loaded models to drastically decline in performance over
time.

This isn't the only option, but my intuition is that the model
will get stuck in a rut if it loses its memory. Best to just
save the state of the replay buffer, IMO. YMMV.